+++ app/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
           xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
-        android:name=".App"
+        android:name=".app.App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_my_foursquare"
         android:label="@string/app_name"
