+++ app/src/main/java/com/github/kulebin/myfoursquareapp/MainActivity.java
@@ -7,6 +7,9 @@ import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.Toast;
 
+import com.github.kulebin.myfoursquareapp.common.IView;
+import com.github.kulebin.myfoursquareapp.presenter.VenueListPresenter;
+
 public class MainActivity extends AppCompatActivity implements IView {
 
     private final VenueListPresenter mListPresenter = new VenueListPresenter(this);
