+++ app/src/main/java/com/github/kulebin/myfoursquareapp/app/ContextHolder.java
@@ -1,4 +1,4 @@
-package com.github.kulebin.myfoursquareapp;
+package com.github.kulebin.myfoursquareapp.app;
 
 import android.content.Context;
