+++ app/src/main/java/com/github/kulebin/myfoursquareapp/common/IView.java
@@ -1,4 +1,4 @@
-package com.github.kulebin.myfoursquareapp;
+package com.github.kulebin.myfoursquareapp.common;
 
 public interface IView {
