+++ app/src/main/java/com/github/kulebin/myfoursquareapp/model/Venue.java
@@ -1,4 +1,4 @@
-package com.github.kulebin.myfoursquareapp;
+package com.github.kulebin.myfoursquareapp.model;
 
 public class Venue {
