+++ app/src/main/java/com/github/kulebin/myfoursquareapp/view/VenueDisplayData.java
@@ -1,4 +1,6 @@
-package com.github.kulebin.myfoursquareapp;
+package com.github.kulebin.myfoursquareapp.view;
+
+import com.github.kulebin.myfoursquareapp.model.Venue;
 
 public class VenueDisplayData {
