+++ app/src/main/java/com/github/kulebin/myfoursquareapp/view/VenueItemView.java
@@ -1,4 +1,4 @@
-package com.github.kulebin.myfoursquareapp;
+package com.github.kulebin.myfoursquareapp.view;
 
 public interface VenueItemView {
